@@ -5,58 +5,62 @@
   "scripts": {
     "test": "jest --watch",
     "testAll": "jest --watchAll",
-    "start": "webpack-dev-server",
+    "watch": "webpack -w",
+    "dev": "webpack serve",
     "build": "webpack"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@types/dat.gui": "^0.7.2",
-    "@types/ramda": "^0.26.33",
-    "@types/react": "^16.8.4",
-    "@types/react-dom": "^16.8.2",
-    "@types/react-transition-group": "^4.2.3",
-    "@types/stats.js": "^0.17.0",
-    "@types/three": "^0.103.2",
     "dat.gui": "^0.7.5",
-    "gsap": "^3.0.1",
-    "ramda": "^0.26.1",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3",
+    "gsap": "^3.12.2",
+    "ramda": "^0.29.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-transition-group": "^4.3.0",
     "stats.js": "^0.17.0",
-    "three": "^0.110.0"
+    "three": "^0.155.0"
   },
   "devDependencies": {
+    "@types/dat.gui": "^0.7.2",
     "@types/jest": "^24.0.22",
-    "@types/jsdom": "^12.2.2",
-    "@types/react-test-renderer": "^16.8.1",
-    "@types/webpack": "^4.4.24",
+    "@types/jsdom": "^21.1.1",
+    "@types/node": "^20.5.0",
+    "@types/ramda": "^0.29.3",
+    "@types/react": "^18.2.20",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-transition-group": "^4.2.3",
+    "@types/stats.js": "^0.17.0",
+    "@types/three": "^0.155.0",
+    "@types/webpack": "^5.28.1",
     "autoprefixer": "^9.4.8",
     "canvas": "^2.3.1",
-    "css-loader": "^3.2.0",
+    "css-loader": "^6.8.1",
     "cssnano": "^4.1.10",
-    "file-loader": "^4.2.0",
+    "eslint": "^8.47.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "file-loader": "^6.2.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^24.9.0",
     "jest-canvas-mock": "^2.2.0",
-    "jsdom": "^15.2.1",
-    "mini-css-extract-plugin": "^0.8.0",
-    "node-sass": "^4.11.0",
-    "postcss": "^7.0.14",
-    "postcss-loader": "^3.0.0",
-    "react-test-renderer": "^16.8.3",
-    "sass-loader": "^8.0.0",
-    "style-loader": "^1.0.0",
+    "jsdom": "^22.1.0",
+    "mini-css-extract-plugin": "^2.7.6",
+    "postcss": "^8.4.28",
+    "postcss-loader": "^7.3.3",
+    "prettier": "^3.0.2",
+    "sass": "^1.65.1",
+    "sass-loader": "^13.3.2",
+    "style-loader": "^3.3.3",
     "ts-jest": "^24.1.0",
-    "ts-loader": "^6.2.1",
-    "tslint": "^5.13.0",
-    "typescript": "^3.3.3333",
-    "webpack": "^4.29.5",
-    "webpack-bundle-analyzer": "^3.0.4",
-    "webpack-cli": "^3.2.3",
-    "webpack-dev-server": "^3.2.0",
+    "ts-loader": "^9.4.4",
+    "typescript": "^5.1.6",
+    "webpack": "^5.88.2",
+    "webpack-bundle-analyzer": "^4.9.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
     "webpack-glsl-loader": "^1.0.1"
   },
   "description": ""
