@@ -1,6 +1,6 @@
-import React, { Component } from "react";
-import Menu from "./Menu";
-import Game from "./Game";
+import React, { Component } from 'react';
+import Menu from './Menu';
+import Game from './Game';
 
 export default class MainApp extends Component {
     // public state = {
