@@ -1,6 +1,6 @@
-import { TimelineLite, TweenLite } from "gsap";
-import Animation from "./Animation";
-import { app } from ".";
+import { gsap } from 'gsap';
+import Animation from './Animation';
+import { app } from '.';
 
 /**
  * Curve
@@ -8,8 +8,14 @@ import { app } from ".";
 export function curveToStep(step: string) {
     const { canvas, curve } = Animation.canvasComponents;
 
-    return TweenLite.to(curve, 0.5, {
-        origin: curve.resizeOptions[step](canvas.width, canvas.height, app.isMobileDevice, app.state.faction),
+    return gsap.to(curve, {
+        duration: 0.5,
+        origin: curve.resizeOptions[step](
+            canvas.width,
+            canvas.height,
+            app.isMobileDevice,
+            app.state.faction,
+        ),
         onComplete: Animation.setWaveInDefaultMode,
     });
 }
@@ -19,9 +25,15 @@ export function curveToStep(step: string) {
  */
 export function lightToStep(step: string) {
     const { canvas, light } = Animation.canvasComponents;
-    const coordinate = light.resizeOptions[step](canvas.width, canvas.height, app.isMobileDevice, app.state.faction);
+    const coordinate = light.resizeOptions[step](
+        canvas.width,
+        canvas.height,
+        app.isMobileDevice,
+        app.state.faction,
+    );
 
-    return TweenLite.to(light, 0.5, {
+    return gsap.to(light, {
+        duration: 0.5,
         delay: 0.1,
         startX: coordinate.x,
         startY: coordinate.y,
@@ -33,9 +45,15 @@ export function lightToStep(step: string) {
  */
 export function shadowToStep(step: string) {
     const { canvas, shadow } = Animation.canvasComponents;
-    const coordinate = shadow.resizeOptions[step](canvas.width, canvas.height, app.isMobileDevice, app.state.faction);
+    const coordinate = shadow.resizeOptions[step](
+        canvas.width,
+        canvas.height,
+        app.isMobileDevice,
+        app.state.faction,
+    );
 
-    return TweenLite.to(shadow, 0.5, {
+    return gsap.to(shadow, {
+        duration: 0.5,
         delay: 0.1,
         startX: coordinate.x,
         startY: coordinate.y,
@@ -47,9 +65,11 @@ export function shadowToStep(step: string) {
  */
 export function homeOut() {
     const { mainTitle, subtitleHome } = Animation.canvasComponents;
-    const homeInterface = Animation.components.homeInterface.current as HTMLDivElement;
+    const homeInterface = Animation.components.homeInterface
+        .current as HTMLDivElement;
 
-    return TweenLite.to([mainTitle, subtitleHome, homeInterface], 0.5, {
+    return gsap.to([mainTitle, subtitleHome, homeInterface], {
+        duration: 0.5,
         opacity: 0,
         onComplete: () => {
             mainTitle.onTransition = false;
@@ -58,22 +78,23 @@ export function homeOut() {
             subtitleHome.onTransition = false;
             subtitleHome.isMount = false;
 
-            homeInterface.style.display = "none";
+            homeInterface.style.display = 'none';
         },
     });
 }
 
 export function homeIn() {
     const { mainTitle, subtitleHome } = Animation.canvasComponents;
-    const homeInterface = Animation.components.homeInterface.current as HTMLDivElement;
+    const homeInterface = Animation.components.homeInterface
+        .current as HTMLDivElement;
 
-    return TweenLite.to([mainTitle, subtitleHome, homeInterface], 0.5, {
+    return gsap.to([mainTitle, subtitleHome, homeInterface], 0.5, {
         opacity: 1,
         onStart: () => {
             mainTitle.onTransition = true;
             subtitleHome.onTransition = true;
 
-            homeInterface.style.display = "block";
+            homeInterface.style.display = 'block';
         },
         onComplete: () => {
             mainTitle.onTransition = false;
@@ -89,23 +110,27 @@ export function homeIn() {
  * Level
  */
 export function levelOut() {
-    const levelInterface = Animation.components.levelInterface.current as HTMLElement;
+    const levelInterface = Animation.components.levelInterface
+        .current as HTMLElement;
 
-    return TweenLite.to(levelInterface, 0.5, {
+    return gsap.to(levelInterface, {
+        duration: 0.5,
         opacity: 0,
         onComplete: () => {
-            levelInterface.style.display = "none";
+            levelInterface.style.display = 'none';
         },
     });
 }
 
 export function levelIn() {
-    const levelInterface = Animation.components.levelInterface.current as HTMLElement;
+    const levelInterface = Animation.components.levelInterface
+        .current as HTMLElement;
 
-    return TweenLite.to(levelInterface, 0.5, {
+    return gsap.to(levelInterface, {
+        duration: 0.5,
         opacity: 1,
         onStart: () => {
-            levelInterface.style.display = "block";
+            levelInterface.style.display = 'block';
         },
     });
 }
@@ -115,9 +140,11 @@ export function levelIn() {
  */
 export function factionOut() {
     const { titleFaction } = Animation.canvasComponents;
-    const factionInterface = Animation.components.factionInterface.current as HTMLElement;
+    const factionInterface = Animation.components.factionInterface
+        .current as HTMLElement;
 
-    return TweenLite.to([titleFaction, factionInterface], 0.5, {
+    return gsap.to([titleFaction, factionInterface], {
+        duration: 0.5,
         opacity: 0,
         onStart: () => {
             titleFaction.onTransition = true;
@@ -125,19 +152,21 @@ export function factionOut() {
         onComplete: () => {
             titleFaction.onTransition = false;
             titleFaction.isMount = false;
-            factionInterface.style.display = "none";
+            factionInterface.style.display = 'none';
         },
     });
 }
 
 export function factionIn() {
     const { titleFaction } = Animation.canvasComponents;
-    const factionInterface = Animation.components.factionInterface.current as HTMLElement;
+    const factionInterface = Animation.components.factionInterface
+        .current as HTMLElement;
 
-    return TweenLite.to([titleFaction, factionInterface], 0.5, {
+    return gsap.to([titleFaction, factionInterface], {
+        duration: 0.5,
         opacity: 1,
         onStart: () => {
-            factionInterface.style.display = "block";
+            factionInterface.style.display = 'block';
             titleFaction.onTransition = true;
         },
         onComplete: () => {
@@ -151,23 +180,27 @@ export function factionIn() {
  * Queue
  */
 export function queueIn() {
-    const queueInterface = Animation.components.queueInterface.current as HTMLElement;
+    const queueInterface = Animation.components.queueInterface
+        .current as HTMLElement;
 
-    return TweenLite.to(queueInterface, 0.5, {
+    return gsap.to(queueInterface, {
+        duration: 0.5,
         opacity: 1,
         onStart: () => {
-            queueInterface.style.display = "block";
+            queueInterface.style.display = 'block';
         },
     });
 }
 
 export function queueOut() {
-    const queueInterface = Animation.components.queueInterface.current as HTMLElement;
+    const queueInterface = Animation.components.queueInterface
+        .current as HTMLElement;
 
-    return TweenLite.to(queueInterface, 0.5, {
+    return gsap.to(queueInterface, {
+        duration: 0.5,
         opacity: 0,
         onComplete: () => {
-            queueInterface.style.display = "none";
+            queueInterface.style.display = 'none';
         },
     });
 }
